+++ src/main/java/scratch/kevin/prvi25/BunceMainRidgeConnectionTests.java
@@ -1,4 +1,4 @@
-package scratch.kevin.nshm23.prvi;
+package scratch.kevin.prvi25;
 
 import java.io.IOException;
 import java.util.List;
