+++ src/main/java/scratch/kevin/prvi25/LowerTraceSubsectionTests.java
@@ -1,4 +1,4 @@
-package scratch.kevin.nshm23.prvi;
+package scratch.kevin.prvi25;
 
 import java.awt.Color;
 import java.awt.geom.Point2D;
