+++ src/main/java/scratch/kevin/prvi25/SubductionDefModConvert.java
@@ -1,4 +1,4 @@
-package scratch.kevin.nshm23.prvi;
+package scratch.kevin.prvi25;
 
 import java.awt.Color;
 import java.awt.Font;
